Add a message template to room.pug with more style
The room has a few extra stuff now, before the message container was empty, but now has a format for messages and placeholder messages. I added some more styling like a scrollbar for messages if it goes offscreen. Additionally, the nickname pug was created as a basic page for now.